+++ src/connectedk8s/azext_connectedk8s/_client_factory.py
@@ -19,13 +19,18 @@ def cf_connected_cluster(cli_ctx, _):
     return cf_connectedk8s(cli_ctx).connected_cluster
 
 
-def cf_connectedk8s_prev_2021_04_01(cli_ctx, *_):
-    from azext_connectedk8s.vendored_sdks.preview_2021_04_01 import ConnectedKubernetesClient
+def cf_connectedk8s_prev_2022_05_01(cli_ctx, *_):
+    from azext_connectedk8s.vendored_sdks.preview_2022_05_01 import ConnectedKubernetesClient
     return get_mgmt_service_client(cli_ctx, ConnectedKubernetesClient)
 
 
-def cf_connected_cluster_prev_2021_04_01(cli_ctx, _):
-    return cf_connectedk8s_prev_2021_04_01(cli_ctx).connected_cluster
+def cf_connected_cluster_prev_2022_05_01(cli_ctx, _):
+    return cf_connectedk8s_prev_2022_05_01(cli_ctx).connected_cluster
+
+
+def cf_connectedmachine(cli_ctx, subscription_id):
+    from azure.mgmt.hybridcompute import HybridComputeManagementClient
+    return get_mgmt_service_client(cli_ctx, HybridComputeManagementClient, subscription_id=subscription_id).private_link_scopes
 
 
 def cf_resource_groups(cli_ctx, subscription_id=None):